requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
asgiref==3.8.1
attrs==23.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Django==5.0.6
djangorestframework==3.15.1
elastic-transport==8.13.0
elasticsearch==8.13.1
fake-useragent==1.5.1
frozenlist==1.4.1
h11==0.14.0
idna==3.7
loguru==0.7.2
lxml==5.2.2
multidict==6.0.5
outcome==1.3.0.post0
pillow==10.3.0
psycopg2-binary==2.9.9
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
PySocks==1.7.1
pyTelegramBotAPI==4.18.1
python-dateutil==2.9.0.post0
requests==2.32.1
schedule==1.2.2
selenium==4.21.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.5.0
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
win32-setctime==1.1.0
wsproto==1.2.0
yarl==1.9.4
